refactor: split tests/ops.rs into 11 modules
