+++ tests/ops/helpers.rs
@@ -1,0 +1,27 @@
+use std::fs;
+use std::path::Path;
+
+use llm_wiki::git;
+
+pub fn setup_wiki(dir: &Path, name: &str) -> std::path::PathBuf {
+    let config_path = dir.join("state").join("config.toml");
+    let wiki_path = dir.join(name);
+
+    llm_wiki::spaces::create(&wiki_path, name, None, false, true, &config_path).unwrap();
+
+    let wiki_root = wiki_path.join("wiki");
+    fs::create_dir_all(wiki_root.join("concepts")).unwrap();
+    fs::write(
+        wiki_root.join("concepts/moe.md"),
+        "---\ntitle: \"MoE\"\ntype: concept\nstatus: active\ntags: [ml]\n---\n\nMixture of Experts.\n",
+    )
+    .unwrap();
+    fs::write(
+        wiki_root.join("concepts/transformer.md"),
+        "---\ntitle: \"Transformer\"\ntype: concept\nstatus: active\n---\n\nAttention is all you need. See [[concepts/moe]].\n",
+    )
+    .unwrap();
+    git::commit(&wiki_path, "add pages").unwrap();
+
+    config_path
+}
