+++ tests/ops/main.rs
@@ -1,0 +1,13 @@
+mod helpers;
+
+mod config;
+mod content;
+mod graph;
+mod history;
+mod hot_reload;
+mod index;
+mod ingest;
+mod schema;
+mod search;
+mod spaces;
+mod watch;
